Prototype: Skip orientation steps as needed
