+++ prototype/src/js/DeviceOrienter.jsx
@@ -1,7 +1,11 @@
 import React from 'react';
 import AlphaOrienter from './AlphaOrienter';
 import BetaOrienter from './BetaOrienter';
-import { delay } from './utils';
+import {
+    delay,
+    calculateNormalizedAlphaOffset,
+    calculateNormalizedBetaOffset
+} from './utils';
 const classNames = require('classnames');
 
 export default class DeviceOrienter extends React.Component {
@@ -9,16 +13,14 @@ export default class DeviceOrienter extends React.Component {
         super(props);
         this.state = {
             supportedEventName: null,
-            betaOriented: false,
+            betaOrientationComplete: false,
             alpha: NaN,
             beta: NaN
         };
     }
 
     componentDidMount() {
-        delay(1000, () => {
-            this.initOrientationListener();
-        });
+        this.initOrientationListener();
     }
 
     componentWillUnmount() {
@@ -86,25 +88,59 @@ export default class DeviceOrienter extends React.Component {
     };
 
     onBetaOrientationComplete = () => {
-        this.setState({ betaOriented: true });
+        this.setState({ betaOrientationComplete: true });
     };
 
-    render() {
+    isBetaOriented() {
+        return calculateNormalizedBetaOffset(this.state.beta) === 0;
+    }
+
+    isAlphaOriented() {
+        return (
+            calculateNormalizedAlphaOffset(
+                this.props.zone.bearing,
+                this.state.alpha
+            ) === 0
+        );
+    }
+
+    determineOrienter() {
         const { alpha, beta } = this.state;
 
-        const orienter = this.state.betaOriented ? (
+        const alphaOrienter = (
             <AlphaOrienter
                 alpha={alpha}
                 target={this.props.zone.bearing}
                 onOrientationComplete={this.onAlphaOrientationComplete}
             />
-        ) : (
+        );
+
+        const betaOrienter = (
             <BetaOrienter
                 beta={beta}
                 onOrientationComplete={this.onBetaOrientationComplete}
             />
         );
 
+        if (!this.state.betaOrientationComplete) {
+            if (this.isBetaOriented()) {
+                this.setState({ betaOrientationComplete: true });
+                return alphaOrienter;
+            } else {
+                return betaOrienter;
+            }
+        } else {
+            if (this.isAlphaOriented()) {
+                this.onAlphaOrientationComplete();
+                return null;
+            } else {
+                return alphaOrienter;
+            }
+        }
+    }
+
+    render() {
+        const orienter = this.determineOrienter();
         return <div className="device-orienter">{orienter}</div>;
     }
 }
